Fix problems related to Mac downloads. For example, if a Mac with OS 10.9 goes to download.php, they see a button for 7.22.2, which only runs on 10.13+.
Solution: in versions.inc, create 'platforms' mac_10_9_plus and max_10_10_plus,
which contain version for those ranges of OS versions.
In download.php, use these depending on the OS version reported in user agent.
In download_all.php, skip them.
This is sort of a kludge, but it's the simplest thing I could think of.

In the main Mac platform, don't use the subtitle 'MacOS 10.10+'; not correct.

Note: to debug download button stuff,
you can manually specify the user agent string as follows:
https://boinc.berkeley.edu/download.php?user_agent=Mozilla/5.0 (Macintosh; Intel Mac OS X 13_5) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.5 Safari/605.1.15
(change the 13_5 to whatever you want).